requirements.txt
streamlit>=1.30.0
markitdown[all]>=0.0.1a
huggingface_hub>=0.20.0
imageio-ffmpeg>=0.4.9
faster-whisper>=0.10.0
yt-dlp>=2025.1.1
pydub>=0.25.1
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.0
xlrd>=2.0.1
PyMuPDF>=1.23.0
extract-msg>=0.45.0
supabase>=2.0.0
python-dotenv>=1.0.0
